Fix file reading & rendering, use correct type for reading

+++ src/loader.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "helpers.h"
 
-EFI_STATUS load_data (UINT8** buffer, UINT64* size, CHAR16* file_name, EFI_HANDLE image);
+EFI_STATUS load_data (UINT32** buffer, UINT64* size, CHAR16* file_name, EFI_HANDLE image);
